backend-python/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
firebase-admin
google-generativeai
httpx
requests
python-multipart
